requirements.txt
pandas
yagmail
supabase
python-dotenv
